fixed yet another gc bug

+++ src/value.c
@@ -33,7 +33,7 @@ const int copy_value(struct value* dest, struct value* src) {
 	return dest->type != value_type_object;
 }
 
-const int compare_value(struct value* a, struct value* b) {
+const double compare_value(struct value* a, struct value* b) {
 	if (a->type != b->type)
 		return a->type - b->type;
